requirements.txt
flask
torch
torchvision
facenet-pytorch
Pillow
numpy